requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0
pillow==10.2.0
pytesseract==0.3.10
openai==1.12.0
